Transformer/requirements.txt
async-timeout==4.0.2
certifi==2022.9.24
cffi==1.15.1
crc16==0.1.1
cryptography==38.0.3
dkimpy==1.0.5
dnspython==2.2.1
packaging==21.3
pycparser==2.21
pyparsing==3.0.9
PyYAML==6.0
redis==4.3.5
reliable-queue==0.0.2
